fastapi-app/requirements.txt
fastapi==0.116.1
requests==2.32.5
alembic==1.16.5
psycopg2-binary==2.9.10
SQLAlchemy==2.0.37
python-dotenv==1.0.1